daemon: move snap command check to notices, tweak error message
