+++ daemon/api_notices.go
@@ -21,7 +21,9 @@ import (
 	"fmt"
 	"math"
 	"net/http"
+	"path/filepath"
 	"strconv"
+	"strings"
 
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/overlord/auth"
@@ -300,11 +302,54 @@ func (inst *noticeInstruction) validate(r *http.Request) *apiError {
 	}
 }
 
+// isRequestFromSnapCmd checks that the request is coming from snap command.
+//
+// It checks that the request process "/proc/PID/exe" points to one of the
+// known locations of the snap command. This not a security-oriented check.
+func isRequestFromSnapCmd(r *http.Request) (bool, error) {
+	ucred, err := ucrednetGet(r.RemoteAddr)
+	if err != nil {
+		return false, err
+	}
+	exe, err := osReadlink(fmt.Sprintf("/proc/%d/exe", ucred.Pid))
+	if err != nil {
+		return false, err
+	}
+
+	// There aren't too many options, but overall possibilities are:
+	// - we are re-executed and the client isn't
+	// - the client re-executed but we did not
+	// - TODO the client could be a merged snapd-snap binary
+
+	switch filepath.Base(exe) {
+	case "snap", "snap-fips": // the standalone snap binary, or its FIPS build variant
+	default:
+		return false, nil
+	}
+
+	if strings.HasPrefix(exe, filepath.Join(dirs.SnapMountDir, "snapd")+"/") ||
+		strings.HasPrefix(exe, filepath.Join(dirs.SnapMountDir, "core")+"/") {
+		// client with expected name from snap or core snap
+		return true, nil
+	}
+
+	if strings.HasPrefix(exe, filepath.Join(dirs.GlobalRootDir, "usr/bin")+"/") {
+		// client with expected name from one of the system locations
+		return true, nil
+	}
+
+	return false, nil
+}
+
 func (inst *noticeInstruction) validateSnapRunInhibitNotice(r *http.Request) *apiError {
+	// double check that the request comes from snap so we can produce a
+	// reasonable error for misguided requests to this currently non-public API.
+	// Note this is not a security check, but merely a low key effort to prevent
+	// misuse of the API.
 	if fromSnapCmd, err := isRequestFromSnapCmd(r); err != nil {
 		return InternalError("cannot check request source: %v", err)
 	} else if !fromSnapCmd {
-		return Forbidden("only snap command can record notices")
+		return Forbidden(`cannot add notice (can only record %q notices from the "snap" command)`, state.SnapRunInhibitNotice)
 	}
 
 	if err := naming.ValidateInstance(inst.Key); err != nil {
