+++ daemon/api_notices_test.go
@@ -1013,7 +1013,7 @@ func (s *noticesSuite) testAddNoticesSnapCmd(c *C, exePath string, shouldFail bo
 	if shouldFail {
 		rsp := s.errorReq(c, req, nil, actionIsExpected)
 		c.Check(rsp.Status, Equals, 403)
-		c.Assert(rsp.Message, Matches, "only snap command can record notices")
+		c.Assert(rsp.Message, Matches, `cannot add notice \(can only record "snap-run-inhibit" notices from the "snap" command\)`)
 	} else {
 		rsp := s.syncReq(c, req, nil, actionIsExpected)
 		c.Assert(rsp.Status, Equals, 200)
@@ -1219,4 +1219,40 @@ func noticeToMap(c *C, notice *state.Notice) map[string]any {
 func addNotice(c *C, st *state.State, userID *uint32, noticeType state.NoticeType, key string, options *state.AddNoticeOptions) {
 	_, err := st.AddNotice(userID, noticeType, key, options)
 	c.Assert(err, IsNil)
+}
+
+func (s *noticesSuite) TestIsFromSnapCmd(c *C) {
+	req, err := http.NewRequest("GET", "/v2/system-volumes", nil)
+	c.Assert(err, IsNil)
+
+	restore := daemon.MockUcrednetGet(func(remoteAddr string) (ucred *daemon.Ucrednet, err error) {
+		return &daemon.Ucrednet{Uid: 42, Pid: 100, Socket: dirs.SnapSocket}, nil
+	})
+	defer restore()
+
+	for _, tc := range []struct {
+		exe string
+		res bool
+	}{
+		{filepath.Join(dirs.SnapMountDir, "core/123/usr/bin/snap"), true},
+		{filepath.Join(dirs.SnapMountDir, "snapd/123/usr/bin/snap"), true},
+		{filepath.Join(dirs.SnapMountDir, "snapd/123/usr/bin/snap-fips"), true},
+		{filepath.Join(dirs.GlobalRootDir, "/usr/bin/snap"), true},
+
+		{filepath.Join(dirs.GlobalRootDir, "/foo/bar/baz/snap"), false},
+		{filepath.Join(dirs.GlobalRootDir, "/foo/bar/baz/not-a-snap"), false},
+	} {
+		c.Logf("tc: %+v", tc)
+		func() {
+			restore = daemon.MockOsReadlink(func(p string) (string, error) {
+				c.Check(p, Equals, "/proc/100/exe")
+				return tc.exe, nil
+			})
+			defer restore()
+
+			res, err := daemon.IsRequestFromSnapCmd(req)
+			c.Check(err, IsNil)
+			c.Check(res, Equals, tc.res)
+		}()
+	}
 }
